refactor: unify implementation
ref #2318
