+++ visualization/app/codeCharta/ui/customConfigs/uploadCustomConfigButton/uploadCustomConfigButton.component.ts
@@ -1,6 +1,8 @@
 import { Component, Inject } from "@angular/core"
 import { Store } from "../../../state/angular-redux/store"
-import { UploadFilesEffect } from "../../../state/effects/uploadFiles/uploadFiles.effect"
+import { CustomConfigHelper } from "../../../util/customConfigHelper"
+import { createCCFileInput } from "../../../util/uploadFiles/createCCFileInput"
+import { readFiles } from "../../../util/uploadFiles/readFiles"
 
 @Component({
 	selector: "cc-upload-custom-config-button",
@@ -10,6 +12,17 @@ export class UploadCustomConfigButtonComponent {
 	constructor(@Inject(Store) private store: Store) {}
 
 	upload() {
-		this.store.dispatch({ type: UploadFilesEffect.uploadCustomConfigsActionType })
+		const fileInput = createCCFileInput()
+		fileInput.addEventListener("change", async () => {
+			const customConfigsContent = await Promise.all(readFiles(fileInput.files))
+			for (const customConfigContent of customConfigsContent) {
+				try {
+					CustomConfigHelper.importCustomConfigs(customConfigContent)
+				} catch {
+					// Explicitly ignored
+				}
+			}
+		})
+		fileInput.click()
 	}
 }
